Reorganized resource builder and users resource files
